+++ vite.config.ts
@@ -13,6 +13,7 @@ export default defineConfig({
       input: {
         index: resolve(__dirname, "index.html"),
         popover: resolve(__dirname, "popover.html"),
+        background: resolve(__dirname, "background.html"),
       },
     },
   },